lossless-claw-be8: extract engine message normalization
