+++ src/engine.ts
@@ -18,7 +18,6 @@ import type {
   SubagentSpawnPreparation,
 } from "./openclaw-bridge.js";
 import {
-  blockFromPart,
   contentFromParts,
   ContextAssembler,
   pickToolCallId,
@@ -65,14 +64,24 @@ import {
 import {
   ConversationStore,
   type ConversationRecord,
-  type CreateMessagePartInput,
-  type MessagePartRecord,
-  type MessagePartType,
 } from "./store/conversation-store.js";
 import { SummaryStore } from "./store/summary-store.js";
 import { createLcmSummarizeFromLegacyParams, LcmProviderAuthError } from "./summarize.js";
 import type { LcmDependencies, StartupSessionFileCandidate } from "./types.js";
 import { estimateTokens } from "./estimate-tokens.js";
+import {
+  RAW_PAYLOAD_EXTERNALIZATION_REASON,
+  buildMessageParts,
+  createBootstrapEntryHash,
+  estimateSessionTokenCountForAfterTurn,
+  extractStructuredText,
+  filterPersistableMessages,
+  hasPersistableMessageRole,
+  hasReplayCriticalRawBlock,
+  serializeRawPayloadContent,
+  toStoredMessage,
+  type StoredMessage,
+} from "./engine/message-normalization.js";
 import { createLcmDatabaseBackup } from "./plugin/lcm-db-backup.js";
 import {
   DatabaseTransactionTimeoutError,
@@ -471,743 +480,6 @@ function normalizeRotateTailMessageCount(value: number, branchMessageCount: numb
     return 1;
   }
   return Math.max(1, Math.min(branchMessageCount, Math.floor(value)));
-}
-
-function appendTextValue(value: unknown, out: string[]): void {
-  if (typeof value === "string") {
-    out.push(value);
-    return;
-  }
-  if (Array.isArray(value)) {
-    for (const entry of value) {
-      appendTextValue(entry, out);
-    }
-    return;
-  }
-  if (!value || typeof value !== "object") {
-    return;
-  }
-
-  const record = value as Record<string, unknown>;
-  appendTextValue(record.text, out);
-  appendTextValue(record.value, out);
-}
-
-const STRUCTURED_TEXT_FIELD_KEYS = ["text", "transcript", "transcription", "message", "summary"];
-const STRUCTURED_ARRAY_FIELD_KEYS = [
-  "segments",
-  "utterances",
-  "paragraphs",
-  "alternatives",
-  "words",
-  "items",
-  "results",
-];
-const STRUCTURED_NESTED_FIELD_KEYS = ["content", "output", "result", "payload", "data", "value"];
-const MAX_STRUCTURED_TEXT_DEPTH = 6;
-const TOOL_RAW_TYPES: ReadonlySet<string> = new Set([
-  "tool_use",
-  "toolUse",
-  "tool-use",
-  "toolCall",
-  "tool_call",
-  "functionCall",
-  "function_call",
-  "function_call_output",
-  "tool_result",
-  "toolResult",
-  "tool_use_result",
-]);
-const REPLAY_CRITICAL_RAW_TYPES: ReadonlySet<string> = new Set([
-  ...TOOL_RAW_TYPES,
-  "thinking",
-  "reasoning",
-]);
-const RAW_PAYLOAD_EXTERNALIZATION_REASON = "large_raw_message";
-
-function looksLikeJsonPayload(value: string): boolean {
-  if (typeof value !== "string") return false;
-  const trimmed = value.trim();
-  if (!trimmed) {
-    return false;
-  }
-  return (
-    (trimmed.startsWith("{") && trimmed.endsWith("}")) ||
-    (trimmed.startsWith("[") && trimmed.endsWith("]"))
-  );
-}
-
-function extractStructuredText(value: unknown, depth: number = 0): string | undefined {
-  if (value == null || depth > MAX_STRUCTURED_TEXT_DEPTH) {
-    return undefined;
-  }
-  if (typeof value === "string") {
-    if (looksLikeJsonPayload(value)) {
-      try {
-        const parsed = JSON.parse(value.trim());
-        const parsedText = extractStructuredText(parsed, depth + 1);
-        if (typeof parsedText === "string" && parsedText.length > 0) {
-          return parsedText;
-        }
-      } catch {
-        // Fall through to returning the original string when parsing fails.
-      }
-    }
-    return value;
-  }
-  if (Array.isArray(value)) {
-    const texts: string[] = [];
-    for (const entry of value) {
-      const text = extractStructuredText(entry, depth + 1);
-      if (typeof text === "string" && text.trim().length > 0) {
-        texts.push(text);
-      }
-    }
-    return texts.length > 0 ? texts.join("\n") : undefined;
-  }
-  if (typeof value !== "object") {
-    return undefined;
-  }
-
-  const record = value as Record<string, unknown>;
-
-  // Skip tool call/result objects — their structured data belongs in the parts table, not content
-  if (typeof record.type === "string" && TOOL_RAW_TYPES.has(record.type)) {
-    if (safeBoolean(record.toolOutputExternalized)) {
-      const externalizedText =
-        extractStructuredText(record.output, depth + 1) ??
-        extractStructuredText(record.content, depth + 1) ??
-        extractStructuredText(record.result, depth + 1);
-      if (typeof externalizedText === "string" && externalizedText.trim().length > 0) {
-        return externalizedText;
-      }
-    }
-    return undefined;
-  }
-
-  for (const key of STRUCTURED_TEXT_FIELD_KEYS) {
-    const candidate = record[key];
-    if (typeof candidate === "string" && candidate.trim().length > 0) {
-      return candidate;
-    }
-  }
-
-  for (const key of STRUCTURED_ARRAY_FIELD_KEYS) {
-    const candidate = record[key];
-    if (Array.isArray(candidate)) {
-      const texts: string[] = [];
-      for (const entry of candidate) {
-        const text = extractStructuredText(entry, depth + 1);
-        if (typeof text === "string" && text.trim().length > 0) {
-          texts.push(text);
-        }
-      }
-      if (texts.length > 0) {
-        return texts.join("\n");
-      }
-    }
-  }
-
-  for (const key of STRUCTURED_NESTED_FIELD_KEYS) {
-    const nested = record[key];
-    const nestedText = extractStructuredText(nested, depth + 1);
-    if (typeof nestedText === "string" && nestedText.trim().length > 0) {
-      return nestedText;
-    }
-  }
-
-  return undefined;
-}
-
-function extractReasoningText(record: Record<string, unknown>): string | undefined {
-  const chunks: string[] = [];
-  appendTextValue(record.summary, chunks);
-  if (chunks.length === 0) {
-    return undefined;
-  }
-
-  const normalized = chunks
-    .map((chunk) => chunk.trim())
-    .filter((chunk, idx, arr) => chunk.length > 0 && arr.indexOf(chunk) === idx);
-  return normalized.length > 0 ? normalized.join("\n") : undefined;
-}
-
-/** Return true when a raw block should remain structurally replayable. */
-function hasReplayCriticalRawBlock(value: unknown): boolean {
-  if (!value || typeof value !== "object") {
-    return false;
-  }
-  if (Array.isArray(value)) {
-    return value.some((entry) => hasReplayCriticalRawBlock(entry));
-  }
-
-  const record = value as Record<string, unknown>;
-  const rawType = safeString(record.type) ?? safeString(record.rawType);
-  if (rawType && REPLAY_CRITICAL_RAW_TYPES.has(rawType)) {
-    return true;
-  }
-
-  for (const key of STRUCTURED_NESTED_FIELD_KEYS) {
-    if (hasReplayCriticalRawBlock(record[key])) {
-      return true;
-    }
-  }
-  for (const key of STRUCTURED_ARRAY_FIELD_KEYS) {
-    if (hasReplayCriticalRawBlock(record[key])) {
-      return true;
-    }
-  }
-
-  return false;
-}
-
-/** Serialize the original message content that backs a generic raw-payload reference. */
-function serializeRawPayloadContent(message: AgentMessage, fallbackContent: string): {
-  content: string;
-  mimeType: string;
-} | null {
-  if (!("content" in message)) {
-    return null;
-  }
-  if (typeof message.content === "string") {
-    return {
-      content: message.content,
-      mimeType: "text/plain",
-    };
-  }
-
-  const serialized = JSON.stringify(message.content);
-  if (typeof serialized !== "string") {
-    return null;
-  }
-  return {
-    content: serialized || fallbackContent,
-    mimeType: "application/json",
-  };
-}
-
-function normalizeUnknownBlock(value: unknown): {
-  type: string;
-  text?: string;
-  metadata: Record<string, unknown>;
-} {
-  if (!value || typeof value !== "object" || Array.isArray(value)) {
-    return {
-      type: "agent",
-      metadata: { raw: value },
-    };
-  }
-
-  const record = value as Record<string, unknown>;
-  const rawType = safeString(record.type);
-  return {
-    type: rawType ?? "agent",
-    text:
-      safeString(record.text) ??
-      safeString(record.thinking) ??
-      ((rawType === "reasoning" || rawType === "thinking")
-        ? extractReasoningText(record)
-        : undefined),
-    metadata: { raw: record },
-  };
-}
-
-function toPartType(type: string): MessagePartType {
-  switch (type) {
-    case "text":
-      return "text";
-    case "thinking":
-    case "reasoning":
-      return "reasoning";
-    case "tool_use":
-    case "toolUse":
-    case "tool-use":
-    case "toolCall":
-    case "functionCall":
-    case "function_call":
-    case "function_call_output":
-    case "tool_result":
-    case "toolResult":
-    case "tool":
-      return "tool";
-    case "patch":
-      return "patch";
-    case "file":
-    case "image":
-      return "file";
-    case "subtask":
-      return "subtask";
-    case "compaction":
-      return "compaction";
-    case "step_start":
-    case "step-start":
-      return "step_start";
-    case "step_finish":
-    case "step-finish":
-      return "step_finish";
-    case "snapshot":
-      return "snapshot";
-    case "retry":
-      return "retry";
-    case "agent":
-      return "agent";
-    default:
-      return "agent";
-  }
-}
-
-/**
- * Convert AgentMessage content into plain text for DB storage.
- *
- * For content block arrays we keep only text blocks to avoid persisting raw
- * JSON syntax that can later pollute assembled model context.
- */
-function extractMessageContent(content: unknown): string {
-  const extracted = extractStructuredText(content);
-  if (typeof extracted === "string") {
-    return extracted;
-  }
-  if (content == null) {
-    return "";
-  }
-  if (Array.isArray(content) && content.length === 0) {
-    return "";
-  }
-  // If content is an array of only tool call/result objects, store as empty
-  // (structured data is preserved in the message parts table)
-  if (Array.isArray(content) && content.length > 0 && content.every(
-    (item) => typeof item === "object" && item !== null && !Array.isArray(item) &&
-      typeof (item as Record<string, unknown>).type === "string" &&
-      TOOL_RAW_TYPES.has((item as Record<string, unknown>).type as string)
-  )) {
-    return "";
-  }
-
-  const serialized = JSON.stringify(content);
-  return typeof serialized === "string" ? serialized : "";
-}
-
-function toRuntimeRoleForTokenEstimate(role: string): "user" | "assistant" | "toolResult" {
-  if (role === "tool" || role === "toolResult") {
-    return "toolResult";
-  }
-  if (role === "user" || role === "system") {
-    return "user";
-  }
-  return "assistant";
-}
-
-function isTextBlock(value: unknown): value is { type: "text"; text: string } {
-  if (!value || typeof value !== "object" || Array.isArray(value)) {
-    return false;
-  }
-  const record = value as Record<string, unknown>;
-  return record.type === "text" && typeof record.text === "string";
-}
-
-function toSyntheticMessagePartRecord(
-  part: CreateMessagePartInput,
-  messageId: number,
-): MessagePartRecord {
-  return {
-    partId: `estimate-part-${part.ordinal}`,
-    messageId,
-    sessionId: part.sessionId,
-    partType: part.partType,
-    ordinal: part.ordinal,
-    textContent: part.textContent ?? null,
-    toolCallId: part.toolCallId ?? null,
-    toolName: part.toolName ?? null,
-    toolInput: part.toolInput ?? null,
-    toolOutput: part.toolOutput ?? null,
-    metadata: part.metadata ?? null,
-  };
-}
-
-function normalizeMessageContentForStorage(params: {
-  message: AgentMessage;
-  fallbackContent: string;
-}): unknown {
-  const { message, fallbackContent } = params;
-  if (!("content" in message)) {
-    return fallbackContent;
-  }
-
-  const role = toRuntimeRoleForTokenEstimate(message.role);
-  const parts = buildMessageParts({
-    sessionId: "storage-estimate",
-    message,
-    fallbackContent,
-  }).map((part) => toSyntheticMessagePartRecord(part, 0));
-
-  if (parts.length === 0) {
-    if (role === "assistant") {
-      return fallbackContent ? [{ type: "text", text: fallbackContent }] : [];
-    }
-    if (role === "toolResult") {
-      return [{ type: "text", text: fallbackContent }];
-    }
-    return fallbackContent;
-  }
-
-  const blocks = parts.map(blockFromPart);
-  if (role === "user" && blocks.length === 1 && isTextBlock(blocks[0])) {
-    return blocks[0].text;
-  }
-  return blocks;
-}
-
-/**
- * Estimate token usage for the content shape that the assembler will emit.
- *
- * LCM stores a plain-text fallback copy in messages.content, but message_parts
- * can rehydrate larger structured/raw blocks. This estimator mirrors the
- * rehydrated shape so compaction decisions use realistic token totals.
- */
-function estimateContentTokensForRole(params: {
-  role: "user" | "assistant" | "toolResult";
-  content: unknown;
-  fallbackContent: string;
-}): number {
-  const { role, content, fallbackContent } = params;
-
-  if (typeof content === "string") {
-    return estimateTokens(content);
-  }
-
-  if (Array.isArray(content)) {
-    if (content.length === 0) {
-      return estimateTokens(fallbackContent);
-    }
-
-    if (role === "user" && content.length === 1 && isTextBlock(content[0])) {
-      return estimateTokens(content[0].text);
-    }
-
-    const serialized = JSON.stringify(content);
-    return estimateTokens(typeof serialized === "string" ? serialized : "");
-  }
-
-  if (content && typeof content === "object") {
-    if (role === "user" && isTextBlock(content)) {
-      return estimateTokens(content.text);
-    }
-
-    const serialized = JSON.stringify([content]);
-    return estimateTokens(typeof serialized === "string" ? serialized : "");
-  }
-
-  return estimateTokens(fallbackContent);
-}
-
-function buildMessageParts(params: {
-  sessionId: string;
-  message: AgentMessage;
-  fallbackContent: string;
-}): import("./store/conversation-store.js").CreateMessagePartInput[] {
-  const { sessionId, message, fallbackContent } = params;
-  const role = typeof message.role === "string" ? message.role : "unknown";
-  const topLevel = message as unknown as Record<string, unknown>;
-  const topLevelToolCallId =
-    safeString(topLevel.toolCallId) ??
-    safeString(topLevel.tool_call_id) ??
-    safeString(topLevel.toolUseId) ??
-    safeString(topLevel.tool_use_id) ??
-    safeString(topLevel.call_id) ??
-    safeString(topLevel.id);
-  const topLevelToolName =
-    safeString(topLevel.toolName) ??
-    safeString(topLevel.tool_name);
-  const topLevelIsError =
-    safeBoolean(topLevel.isError) ??
-    safeBoolean(topLevel.is_error);
-  const rawPayloadExternalized = safeBoolean(topLevel.rawPayloadExternalized);
-  const externalizedFileId = safeString(topLevel.externalizedFileId);
-  const originalByteSize =
-    typeof topLevel.originalByteSize === "number"
-      ? topLevel.originalByteSize
-      : undefined;
-  const externalizationReason = safeString(topLevel.externalizationReason);
-
-  // BashExecutionMessage: preserve a synthetic text part so output is round-trippable.
-  if (!("content" in message) && "command" in message && "output" in message) {
-    return [
-      {
-        sessionId,
-        partType: "text",
-        ordinal: 0,
-        textContent: fallbackContent,
-        metadata: toJson({
-          originalRole: role,
-          source: "bash-exec",
-          command: safeString((message as { command?: unknown }).command),
-        }),
-      },
-    ];
-  }
-
-  if (!("content" in message)) {
-    return [
-      {
-        sessionId,
-        partType: "agent",
-        ordinal: 0,
-        textContent: fallbackContent || null,
-        metadata: toJson({
-          originalRole: role,
-          source: "unknown-message-shape",
-          raw: message,
-        }),
-      },
-    ];
-  }
-
-  if (typeof message.content === "string") {
-    return [
-      {
-        sessionId,
-        partType: "text",
-        ordinal: 0,
-        textContent: message.content,
-        metadata: toJson({
-          originalRole: role,
-          toolCallId: topLevelToolCallId,
-          toolName: topLevelToolName,
-          isError: topLevelIsError,
-          rawPayloadExternalized: rawPayloadExternalized || undefined,
-          externalizedFileId,
-          originalByteSize,
-          externalizationReason,
-        }),
-      },
-    ];
-  }
-
-  if (!Array.isArray(message.content)) {
-    return [
-      {
-        sessionId,
-        partType: "agent",
-        ordinal: 0,
-        textContent: fallbackContent || null,
-        metadata: toJson({
-          originalRole: role,
-          source: "non-array-content",
-          raw: message.content,
-        }),
-      },
-    ];
-  }
-
-  const parts: CreateMessagePartInput[] = [];
-  for (let ordinal = 0; ordinal < message.content.length; ordinal++) {
-    const block = normalizeUnknownBlock(message.content[ordinal]);
-    const metadataRecord = block.metadata.raw as Record<string, unknown> | undefined;
-    const rawBlockType = safeString(metadataRecord?.rawType) ?? block.type;
-    const partType = toPartType(rawBlockType);
-    const rawBlock =
-      metadataRecord && rawBlockType !== block.type
-        ? {
-            ...metadataRecord,
-            type: rawBlockType,
-          }
-        : (metadataRecord ?? message.content[ordinal]);
-    const toolCallId =
-      safeString(metadataRecord?.toolCallId) ??
-      safeString(metadataRecord?.tool_call_id) ??
-      safeString(metadataRecord?.toolUseId) ??
-      safeString(metadataRecord?.tool_use_id) ??
-      safeString(metadataRecord?.call_id) ??
-      (partType === "tool" ? safeString(metadataRecord?.id) : undefined) ??
-      topLevelToolCallId;
-
-    parts.push({
-      sessionId,
-      partType,
-      ordinal,
-      textContent: block.text ?? null,
-      toolCallId,
-      toolName:
-        safeString(metadataRecord?.name) ??
-        safeString(metadataRecord?.toolName) ??
-        safeString(metadataRecord?.tool_name) ??
-        topLevelToolName,
-      toolInput:
-        metadataRecord?.input !== undefined
-          ? toJson(metadataRecord.input)
-          : metadataRecord?.arguments !== undefined
-            ? toJson(metadataRecord.arguments)
-          : metadataRecord?.toolInput !== undefined
-            ? toJson(metadataRecord.toolInput)
-            : (safeString(metadataRecord?.tool_input) ?? null),
-      toolOutput:
-        metadataRecord?.output !== undefined
-          ? toJson(metadataRecord.output)
-          : metadataRecord?.toolOutput !== undefined
-            ? toJson(metadataRecord.toolOutput)
-            : (safeString(metadataRecord?.tool_output) ?? null),
-      metadata: toJson({
-        originalRole: role,
-        toolCallId: topLevelToolCallId,
-        toolName: topLevelToolName,
-        isError: topLevelIsError,
-        externalizedFileId: safeString(metadataRecord?.externalizedFileId),
-        originalByteSize:
-          typeof metadataRecord?.originalByteSize === "number"
-            ? metadataRecord.originalByteSize
-            : undefined,
-        toolOutputExternalized: safeBoolean(metadataRecord?.toolOutputExternalized),
-        externalizationReason: safeString(metadataRecord?.externalizationReason),
-        rawType: rawBlockType,
-        raw: rawBlock,
-      }),
-    });
-  }
-
-  return parts;
-}
-
-/**
- * Map AgentMessage role to the DB enum.
- *
- *   "user"      -> "user"
- *   "assistant" -> "assistant"
- *
- * AgentMessage only has user/assistant roles, but we keep the mapping
- * explicit for clarity and future-proofing.
- */
-function toDbRole(role: string): "user" | "assistant" | "system" | "tool" {
-  if (role === "tool" || role === "toolResult") {
-    return "tool";
-  }
-  if (role === "system") {
-    return "system";
-  }
-  if (role === "user") {
-    return "user";
-  }
-  if (role === "assistant") {
-    return "assistant";
-  }
-  // Direct callers should filter unknown roles before storage. Preserve the
-  // historical fallback for typed AgentMessage values that reach this helper.
-  return "assistant";
-}
-
-function hasPersistableMessageRole(message: AgentMessage): boolean {
-  const role = (message as { role?: unknown }).role;
-  return (
-    role === "user" ||
-    role === "assistant" ||
-    role === "system" ||
-    role === "tool" ||
-    role === "toolResult"
-  );
-}
-
-function filterPersistableMessages(messages: AgentMessage[]): AgentMessage[] {
-  return messages.filter(hasPersistableMessageRole);
-}
-
-type StoredMessage = {
-  role: "user" | "assistant" | "system" | "tool";
-  content: string;
-  tokenCount: number;
-};
-
-/**
- * Normalize AgentMessage variants into the storage shape used by LCM.
- */
-function toStoredMessage(message: AgentMessage): StoredMessage {
-  const content =
-    "content" in message
-      ? extractMessageContent(message.content)
-      : "output" in message
-        ? `$ ${(message as { command: string; output: string }).command}\n${(message as { command: string; output: string }).output}`
-        : "";
-  const runtimeRole = toRuntimeRoleForTokenEstimate(message.role);
-  const normalizedContent =
-    "content" in message
-      ? normalizeMessageContentForStorage({
-          message,
-          fallbackContent: content,
-        })
-      : content;
-  const tokenCount =
-    "content" in message
-      ? estimateContentTokensForRole({
-          role: runtimeRole,
-          content: normalizedContent,
-          fallbackContent: content,
-        })
-      : estimateTokens(content);
-
-  return {
-    role: toDbRole(message.role),
-    content,
-    tokenCount,
-  };
-}
-
-function createBootstrapEntryHash(message: StoredMessage | null): string | null {
-  if (!message) {
-    return null;
-  }
-  return createHash("sha256")
-    .update(JSON.stringify({ role: message.role, content: message.content }))
-    .digest("hex");
-}
-
-function estimateMessageContentTokensForAfterTurn(content: unknown): number {
-  if (typeof content === "string") {
-    return estimateTokens(content);
-  }
-  if (Array.isArray(content)) {
-    let total = 0;
-    for (const part of content) {
-      if (!part || typeof part !== "object") {
-        continue;
-      }
-      const record = part as Record<string, unknown>;
-      const text =
-        typeof record.text === "string"
-          ? record.text
-          : typeof record.thinking === "string"
-            ? record.thinking
-            : "";
-      if (text) {
-        total += estimateTokens(text);
-      }
-    }
-    return total;
-  }
-  if (content == null) {
-    return 0;
-  }
-  const serialized = JSON.stringify(content);
-  return estimateTokens(typeof serialized === "string" ? serialized : "");
-}
-
-function estimateSessionTokenCountForAfterTurn(messages: AgentMessage[]): number {
-  let total = 0;
-  for (const message of messages) {
-    if ("content" in message) {
-      total += estimateMessageContentTokensForAfterTurn(message.content);
-      continue;
-    }
-    if ("command" in message || "output" in message) {
-      const commandText =
-        typeof (message as { command?: unknown }).command === "string"
-          ? (message as { command?: string }).command
-          : "";
-      const outputText =
-        typeof (message as { output?: unknown }).output === "string"
-          ? (message as { output?: string }).output
-          : "";
-      total += estimateTokens(`${commandText}\n${outputText}`);
-    }
-  }
-  return total;
 }
 
 function normalizeNonNegativeInteger(value: unknown): number | undefined {
